Fix parameters : GLuint instead of GLfloat.

+++ DirectionalLight.cpp
@@ -9,8 +9,8 @@ DirectionalLight::DirectionalLight()
 }
 
 DirectionalLight::DirectionalLight(
-	GLfloat shadowWidth,
-	GLfloat shadowHeight,
+	GLuint shadowWidth,
+	GLuint shadowHeight,
 	glm::vec3 colour,
 	GLfloat ambientIntensity,
 	GLfloat diffuseIntensity,
